Separate TGenStep class from BlobDepot and move it to common utils
