+++ ydb/core/util/gen_step.cpp
@@ -1,0 +1,6 @@
+#include "gen_step.h"
+
+template<>
+void Out<NKikimr::TGenStep>(IOutputStream& s, const NKikimr::TGenStep& x) {
+    x.Output(s);
+}
